More reliable rebuilds, refactor symtable code into new file, tests pass
