+++ vm/include/gc.h
@@ -20,15 +20,25 @@ inline uint8 RAM_EXTRACT_TYPE_TAG(uint16 o) {
 
 	return (((f2 & signFill) >> 5) | (signFill & ~(signFill >> 1)) | ((signFill >> 1) << 3)) & 0x0f;
 }
+inline uint8 ROM_EXTRACT_TYPE_TAG(uint16 o) {
+	_Static_assert((uint8_t)((int8_t)0x80 >> 6) == 0xfe, "Platform is not performing sign extension on arithemtic shift?!?");
+	const uint8 f0 = rom_get_field0(o);
+	const uint8 f2 = rom_get_field2(o);
+	// requires sign extension!
+	const uint8 signFill = (uint8_t)(((int8_t)f0) >> 6) /* possible values 0xfe, 0xff, 0x00, 0x01 */;
+
+	return (((f2 & signFill) >> 5) | (signFill & ~(signFill >> 1)) | ((signFill >> 1) << 3)) & 0x0f;
+}
 #else
-#define _ETT_F0(o) (ram_get_field0(o))
-#define _ETT_F2(o) (ram_get_field2(o))
+#define _ETT_F0(aspace, o) (aspace ##_get_field0 (o))
+#define _ETT_F2(aspace, o) (aspace ##_get_field2 (o))
 #define _ETT_SF(f0) ((uint8_t)(((int8_t)f0) >> 6))
 #define _ETT_RHS(f0,f2) ((((uint8_t)f2) & _ETT_SF(f0)) >> 5)
 #define _ETT_LHS_C0(f0,f2) (_ETT_SF(f0) & ~(_ETT_SF(f0) >> 1))
 #define _ETT_LHS_C1(f0,f2) ((_ETT_SF(f0) >> 1) << 3)
 #define _ETT(f0, f2) ((_ETT_RHS(f0, f2) | _ETT_LHS_C0(f0,f2) | _ETT_LHS_C1(f0,f2)) & 0x0f)
-#define RAM_EXTRACT_TYPE_TAG(o) _ETT(_ETT_F0(o), _ETT_F2(o))
+#define RAM_EXTRACT_TYPE_TAG(o) _ETT(_ETT_F0(ram, o), _ETT_F2(ram, o))
+#define ROM_EXTRACT_TYPE_TAG(o) _ETT(_ETT_F0(rom, o), _ETT_F2(rom, o))
 #endif
 
 typedef enum PicobitType {
@@ -83,7 +93,7 @@ uint8 HAS_1_OBJECT_FIELD(uint16 visit)
 extern int max_live;
 #endif
 
-void init_ram_heap (uint8 numConstants);
+void init_ram_heap ();
 
 void mark (obj temp);
 void sweep ();
